# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MICO360 Tasks API
  version: 1.0.0
  description: >-
    REST API for MICO360 Tasks (web app, Chrome extension, and the Phase 2 Android
    app). All responses use a JSON envelope: success is `{ "data": ... }`, errors
    are `{ "error": { "code": string, "message": string, "details"?: any } }`.
    Authenticate by sending `Authorization: Bearer <accessToken>`.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
  - url: https://task.mico360.com/api/v1
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /reports/projects:
    get:
      summary: Project performance report (admin). Also /reports/status, /reports/workload, /reports/completion.
      parameters:
        - { name: projectId, in: query, schema: { type: string }, description: Only this project }
        - { name: userId, in: query, schema: { type: string }, description: Only tasks assigned to this team member }
      responses: { '200': { description: OK }, '403': { description: Forbidden } }
...
  /meetings/{id}/notes/{noteId}/restore:
    post:
      summary: Undo a meeting-note delete (notes are soft-deleted). The note's author or the meeting's organizer only.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: noteId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: The restored note }
        '403': { description: Not the author or organizer }
        '404': { description: No deleted note with that id in this meeting }